[Test] testing remove_TR for nifti + changing prepostprocessing so that nifti is censored properly
